+++ CMakeLists.txt
@@ -31,13 +31,15 @@ add_executable (expirationTracker
   # Entry into SDL after forked from displayEntry
   display/src/sdl2/sdl_entry.cpp
 
-  # GUI
-  display/src/sdl2/element.cpp
-  display/src/sdl2/button.cpp
+  # GUI 
   display/src/sdl2/display.cpp
-  display/src/sdl2/text.cpp
-  display/src/sdl2/panel.cpp
-  display/src/sdl2/scroll_box.cpp
+
+  # GUI Elements
+  display/src/sdl2/elements/element.cpp
+  display/src/sdl2/elements/button.cpp
+  display/src/sdl2/elements/text.cpp
+  display/src/sdl2/elements/panel.cpp
+  display/src/sdl2/elements/scroll_box.cpp
 
   # GUI States
   display/src/sdl2/states/state.cpp
