+++ display/src/sdl2/elements/panel.h
@@ -5,8 +5,8 @@
 #include <memory>
 #include <vector>
 
+#include "../display_global.h"
 #include "button.h"
-#include "display_global.h"
 #include "element.h"
 #include "text.h"
