+++ display/src/sdl2/elements/scroll_box.h
@@ -5,8 +5,8 @@
 #include <memory>
 #include <vector>
 
-#include "../../../food_item.h"
-#include "display_global.h"
+#include "../../../../food_item.h"
+#include "../display_global.h"
 #include "element.h"
 #include "panel.h"
