+++ display/src/sdl2/elements/text.h
@@ -5,7 +5,7 @@
 #include <SDL2/SDL_ttf.h>
 #include <string>
 
-#include "display_global.h"
+#include "../display_global.h"
 #include "element.h"
 
 /**
