+++ display/src/sdl2/states/pause_menu.h
@@ -5,9 +5,9 @@
 #include <memory>
 #include <string>
 
-#include "../button.h"
 #include "../display_global.h"
-#include "../text.h"
+#include "../elements/button.h"
+#include "../elements/text.h"
 #include "state.h"
 
 /**
